Reparing project after failed commit

+++ HospitalManager.App/Managers/PatientManager.cs
@@ -14,7 +14,7 @@ namespace HospitalManager.App.Managers
         {
             _patientService = patientService;
         }
-        public void GetNewPatientData(User user)
+        public void GetNewPatientData(LoginAction user)
         {
             Patient patient = new Patient(user);
             patient.Id = _patientService.Items.Count + 1;
@@ -46,16 +46,13 @@ namespace HospitalManager.App.Managers
 
         public void Remove(int id)
         {
-            var patient = _patientService.GetById(id);
-            _patientService.Remove(patient);
+            foreach (var patient in _patientService.Items)
+            {
+                if (patient.Id == id)
+                {
+                    _patientService.Remove(patient);
+                }
+            }
         }
-
-        public Patient GetById(int id)
-        {
-            var patient = _patientService.GetById(id);
-            return patient;
-        }
-
-
     }
 }
